requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1635819660018/work
argcomplete @ file:///home/conda/feedstock_root/build_artifacts/argcomplete_1641290081297/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
branca==0.4.2
cachetools==4.2.2
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
debugpy @ file:///opt/concourse/worker/volumes/live/839ef258-c5a5-406a-56eb-b19da248f276/volume/debugpy_1629222714909/work
decorator==4.4.2
easydev==0.12.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
faiss==1.6.5
fitter==1.4.0
folium==0.12.1
geographiclib==1.50
geopy @ file:///home/conda/feedstock_root/build_artifacts/geopy_1609078823613/work
idna==2.10
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1639976218118/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1641325382024/work/dist/ipykernel-6.6.1-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1641404859093/work
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1637175421629/work
Jinja2==3.0.1
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1636814373394/work
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.3.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numpy==1.19.4
olefile==0.46
opencv-python==4.4.0.46
pandas==1.2.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
patsy==0.5.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///opt/concourse/worker/volumes/live/95e97a5f-0d90-45c8-4080-207143da6ebf/volume/pillow_1603822285720/work
premailer==3.9.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein @ file:///opt/concourse/worker/volumes/live/82ae8e6d-1325-4a59-7513-def6569b925b/volume/python-levenshtein_1613985211451/work
pytz==2020.5
pyzmq @ file:///opt/concourse/worker/volumes/live/7dec6841-64b1-4d95-7226-ce37c6ca8915/volume/pyzmq_1638435000749/work
requests==2.25.1
scikit-learn @ file:///opt/concourse/worker/volumes/live/4b416fcf-2a1b-4a30-781a-c588a35f6ca3/volume/scikit-learn_1614446663856/work
scipy==1.5.4
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
six @ file:///opt/concourse/worker/volumes/live/f983ba11-c9fe-4dff-7ce7-d89b95b09771/volume/six_1605205318156/work
statsmodels @ file:///opt/concourse/worker/volumes/live/ad2325f1-005b-4882-5896-b1e5f01ee884/volume/statsmodels_1614023733346/work
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado @ file:///opt/concourse/worker/volumes/live/d531d395-893c-4ca1-6a5f-717b318eb08c/volume/tornado_1606942307627/work
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1638334978229/work
urllib3==1.26.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
yagmail==0.14.256
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
